Added `doxygen_runner` rule for running doxygen on a workspace.
