+++ .gitignore
@@ -1,3 +1,4 @@
 # Git ignore patterns
 bazel-*
 user.bazelrc
+docs/rules_doxygen
