+++ doxygen/BUILD.bazel
@@ -1,4 +1,5 @@
 load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
+load(":defs.bzl", "doxygen_runner")
 
 exports_files([
     "doxygen.cfg",
@@ -14,6 +15,10 @@ label_flag(
 toolchain_type(
     name = "toolchain_type",
     visibility = ["//visibility:public"],
+)
+
+doxygen_runner(
+    name = "doxygen",
 )
 
 bzl_library(
