+++ doxygen/private/BUILD.bazel
@@ -2,9 +2,16 @@ load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
 load("@rules_cc//cc:defs.bzl", "cc_binary")
 
 cc_binary(
+    name = "doxygen_action_runner",
+    srcs = ["doxygen_action_runner.cc"],
+    visibility = ["//visibility:public"],
+)
+
+cc_binary(
     name = "doxygen_runner",
     srcs = ["doxygen_runner.cc"],
     visibility = ["//visibility:public"],
+    deps = ["@bazel_tools//tools/cpp/runfiles"],
 )
 
 bzl_library(
